Mise à jour de l'ensembles des fonctions ... on peut passer maintenant à l'encodage multiple
git-svn-id: svn://zone.spip.org/spip-zone/_plugins_@29748 ac52e18a-acf5-0310-9fe8-c4428f23b10a

+++ spipmotion/action/spipmotion_logo.php
@@ -21,7 +21,7 @@ function action_spipmotion_logo_dist(){
 	$arg = $securiser_action();
 
 	if (!preg_match(",^(-?)(\d+)\W(\w+)\W?(\d*)\W?(\d*)$,", $arg, $r)){
-		spip_log("action_infos_video_dist incompris: " . $arg);
+		spip_log("action_logo_video_dist incompris: " . $arg);
 		$redirect = urldecode(_request('redirect'));
 		return;
 	} else action_infos_video_post($r);
@@ -40,30 +40,31 @@ function action_infos_video_post($r){
 	spip_log("on travail sur $chemin","spipmotion");	
 	
 	$movie = new ffmpeg_movie($chemin,0);
-	
 	$string_temp = "$id-$type-$id_document";
 	$query = md5($string_temp);
 	$dossier_temp = _DIR_VAR;
 	$fichier_temp = "$dossier_temp$query.jpg";
 	spip_log("fichier temporaire = $fichier_temp","spipmotion");	
 	
-	$frame1 = $movie->getFrame(400);
-	spip_log($frame1);
-	$img_temp = $frame1->toGDImage();
-	imagejpeg($img_temp, $fichier_temp);
-	$img_finale = $fichier_temp;
-	$mode = 'vignette';
+	$frame1 = $movie->getFrame(100);
+	if($frame1){
+		spip_log('frame1 existe');
+		$img_temp = $frame1->toGDImage();
+		imagejpeg($img_temp, $fichier_temp);
+		$img_finale = $fichier_temp;
+		$mode = 'vignette';
+		
+		$ajouter_documents = charger_fonction('ajouter_documents', 'inc');
 	
-	$ajouter_documents = charger_fonction('ajouter_documents', 'inc');
-
-	spip_log("on ajoute $img","spipmotion");	
-	// verifier l'extension du fichier en fonction de son type mime
-	list($extension,$arg) = fixer_extension_document($arg);
-	$x = $ajouter_documents($img_finale, $img_finale, 
-			    $type, $id, $mode, $id_document, $actifs);
-	
-	imagedestroy($img_temp);
-	unlink($img_finale);
+		spip_log("on ajoute $img","spipmotion");	
+		// verifier l'extension du fichier en fonction de son type mime
+		list($extension,$arg) = fixer_extension_document($arg);
+		$x = $ajouter_documents($img_finale, $img_finale, 
+				    $type, $id, $mode, $id_document, $actifs);
+		
+		imagedestroy($img_temp);
+		unlink($img_finale);
+	}
 	
 	// un invalideur a la hussarde qui doit marcher au moins pour article, breve, rubrique
 	include_spip('inc/invalideur');
